Bombina/requirements.txt
# Bombina Requirements
# Core
ollama>=0.4.0
pyyaml>=6.0

# RAG System
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Fine-tuning (optional - requires GPU)
# unsloth
# trl>=0.7.0
# transformers>=4.36.0
# datasets>=2.14.0
# peft>=0.7.0
# bitsandbytes>=0.41.0
# accelerate>=0.25.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
